requirements.txt
pytest==3.9.2
paramiko==2.4.2
tox==3.5.2
flake8==3.5.0
pytest-cov==2.6.0
coverage==4.5.1
libvirt-python==4.8.0
pexpect==4.6.0